+++ TimeToCook/Modules/CustomTabBar/CustomTabBarController.swift
@@ -1,12 +1,11 @@
 //
 //  CustomTabBarController.swift
-//  VarkaAPP
+//  TimeToCook
 //
-//  Created by Никита Гвоздиков on 05.03.2021.
+//  Created by Никита Гвоздиков on 11.07.2021.
 //
 
 import UIKit
-
 
 protocol AddNewProductViewControllerDelegate: AnyObject {
     func productWasAdded(product: Product?)
