+++ TimeToCook/Modules/Timer/Subviews/TimeDiagram.swift
@@ -1,8 +1,8 @@
 //
 //  TimeDiagram.swift
-//  VarkaAPP
+//  TimeToCook
 //
-//  Created by Никита Гвоздиков on 27.03.2021.
+//  Created by Никита Гвоздиков on 12.08.2021.
 //
 
 import SwiftUI
