+++ TimeToCook/Modules/Timer/Subviews/TimerContentView.swift
@@ -1,0 +1,28 @@
+//
+//  TimerContentView.swift
+//  TimeToCook
+//
+//  Created by Никита Гвоздиков on 10.08.2021.
+//
+
+import UIKit
+
+final class TimerContentView: UIView {
+    
+    //MARK: Init
+    
+    init() {
+        super.init(frame: .zero)
+        translatesAutoresizingMaskIntoConstraints = false
+        backgroundColor = .white
+        layer.shadowRadius = 5
+        layer.shadowOpacity = 0.3
+        layer.shadowOffset = CGSize(width: 0, height: 0)
+        layer.shadowColor = #colorLiteral(red: 0.6666666865, green: 0.6666666865, blue: 0.6666666865, alpha: 1)
+        layer.cornerRadius = UIConstants.defaultCornerRadius
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
