+++ TimeToCook/Modules/Timer/Subviews/TimerStartButton.swift
@@ -1,0 +1,35 @@
+//
+//  StartTimerButton.swift
+//  VarkaAPP
+//
+//  Created by Никита Гвоздиков on 11.08.2021.
+//
+
+import UIKit
+
+final class TimerStartButton: UIButton {
+    
+    //MARK: Init
+    
+    init() {
+        super.init(frame: .zero)
+        clipsToBounds = true
+        titleLabel?.textColor = .white
+        translatesAutoresizingMaskIntoConstraints = false
+        setTitle("Старт", for: .normal)
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    //MARK: Override
+    
+    override func layoutSubviews() {
+        super.layoutSubviews()
+        layer.cornerRadius = layer.frame.height / 2
+        backgroundColor = isEnabled
+            ? UIConstants.startTimerButtonEnabledColor
+            : UIConstants.startTimerButtonDisabledColor
+    }
+}
